chore(cb2-9907): rename ivadefects.json
